+++ src/Accelerators.h
@@ -5,3 +5,4 @@ void FreeAcceleratorTables();
 void CreateSumatraAcceleratorTable();
 HACCEL* GetAcceleratorTables();
 TempStr AppendAccelKeyToMenuStringTemp(TempStr str, int cmdId);
+bool IsValidShortcutString(const char* shortcut);