+++ src/AppSettings.cpp
@@ -96,9 +96,24 @@ static void setMinMax(int& i, int minVal, int maxVal) {
     }
 }
 
+static void SetCommandNameAndShortcut(CustomCommand* cmd, const char* name, const char* key) {
+    if (!cmd) {
+        return;
+    }
+    cmd->name = str::IsEmptyOrWhiteSpace(name) ? nullptr : str::Dup(name);
+    if (str::IsEmptyOrWhiteSpace(key)) {
+        return;
+    }
+    if (!IsValidShortcutString(key)) {
+        logf("SetCommandNameAndShortcut: '%s' is not a valid shortcut for '%s'\n", key, cmd->definition);
+        return;
+    }
+    cmd->key = str::Dup(key);
+}
+
 /* for every selection handler defined by user in advanced settings, create
     a command that will be inserted into a menu item */
-static void CreateUserSelectionHandlerCommands() {
+static void CreateSelectionHandlerCommands() {
     if (!HasPermission(Perm::InternetAccess) || !HasPermission(Perm::CopySelection)) {
         // TODO: when we add exe handlers, only filter the URL ones
         return;
@@ -113,11 +128,31 @@ static void CreateUserSelectionHandlerCommands() {
             continue;
         }
 
-        CommandArg* args = NewStringArg(kCmdArgName, sh->name);
-        CommandArg* arg = NewStringArg(kCmdArgURL, sh->url);
-        InsertArg(&args, arg);
+        CommandArg* args = NewStringArg(kCmdArgURL, sh->url);
         auto cmd = CreateCustomCommand(sh->url, CmdSelectionHandler, args);
-        sh->cmdId = cmd->id;
+        SetCommandNameAndShortcut(cmd, sh->name, sh->key);
+    }
+}
+
+static void CreateExternalViewersCommands() {
+    for (ExternalViewer* ev : *gGlobalPrefs->externalViewers) {
+        if (!ev || str::IsEmptyOrWhiteSpace(ev->commandLine)) {
+            continue;
+        }
+        CommandArg* args = NewStringArg(kCmdArgCommandLine, ev->commandLine);
+        if (!str::IsEmptyOrWhiteSpace(ev->filter)) {
+            auto arg = NewStringArg(kCmdArgFilter, ev->filter);
+            InsertArg(&args, arg);
+        }
+        auto cmd = CreateCustomCommand("", CmdViewWithExternalViewer, args);
+        SetCommandNameAndShortcut(cmd, ev->name, ev->key);
+    }
+}
+
+static void CreateCustomCommands() {
+    for (Shortcut* shortcut : *gGlobalPrefs->shortcuts) {
+        auto [cmdId, cmd] = CreateCommandFromDefinition(shortcut->cmd);
+        SetCommandNameAndShortcut(cmd, shortcut->name, shortcut->key);
     }
 }
 
@@ -218,9 +253,10 @@ bool LoadSettings() {
     // re-create commands
     FreeCustomCommands();
     // Note: some are also created in ReCreateSumatraAcceleratorTable()
-    CreateUserSelectionHandlerCommands();
     CreateThemeCommands();
     CreateExternalViewersCommands();
+    CreateSelectionHandlerCommands();
+    CreateCustomCommands();
 
     // re-create accelerators
     FreeAcceleratorTables();