+++ src/ExternalViewers.cpp
@@ -422,25 +422,6 @@ bool PathMatchFilter(const char* path, const char* filter) {
     return matches;
 }
 
-ExternalViewer* CustomExternalViewerForCmdId(int cmdId) {
-    for (auto& ev : *gGlobalPrefs->externalViewers) {
-        if (ev->cmdId == cmdId) {
-            return ev;
-        }
-    }
-    return nullptr;
-}
-
-void CreateExternalViewersCommands() {
-    for (auto& ev : *gGlobalPrefs->externalViewers) {
-        if (!ev || str::IsEmptyOrWhiteSpace(ev->commandLine)) {
-            continue;
-        }
-        auto cmd = CreateCustomCommand("", CmdViewWithExternalViewer, nullptr);
-        ev->cmdId = cmd->id;
-    }
-}
-
 // TODO: find a better file for this?
 bool RunWithExe(WindowTab* tab, const char* cmdLine, const char* filter) {
     const char* path = tab->filePath;
@@ -475,25 +456,6 @@ bool RunWithExe(WindowTab* tab, const char* cmdLine, const char* filter) {
     }
     TempStr params = JoinTemp(&argsQuoted, " ");
     return LaunchFileShell(exePath, params);
-}
-
-bool ViewWithCustomExternalViewer(WindowTab* tab, int cmdId) {
-    if (!CanAccessDisk() || !tab || !file::Exists(tab->filePath)) {
-        return false;
-    }
-
-    ExternalViewer* ev = nullptr;
-    for (auto& ev2 : *gGlobalPrefs->externalViewers) {
-        if (ev2->cmdId == cmdId) {
-            ev = ev2;
-            break;
-        }
-    }
-    if (!ev) {
-        return false;
-    }
-
-    return RunWithExe(tab, ev->commandLine, ev->filter);
 }
 
 #define DEFINE_GUID_STATIC(name, l, w1, w2, b1, b2, b3, b4, b5, b6, b7, b8) \