+++ src/ExternalViewers.h
@@ -14,7 +14,4 @@ bool SendAsEmailAttachment(WindowTab* tab, HWND hwndParent = nullptr);
 bool CouldBePDFDoc(WindowTab*);
 bool PathMatchFilter(const char* path, const char* filter);
 
-void CreateExternalViewersCommands();
-ExternalViewer* CustomExternalViewerForCmdId(int cmdId);
-bool ViewWithCustomExternalViewer(WindowTab* tab, int cmdId);
 bool RunWithExe(WindowTab* tab, const char* cmdLine, const char* filter);